// linx/utils/types/productJSON.ts
export interface WebPage {
  Response: Response;
  Model: Model;
  PageInfo: PageInfo;
}

export interface Model {
  AcceptanceTermID: number;
  AllowOnlyAdditionalPurchase: boolean;
  Associations: unknown[];
  Availability: string;
  AvailabilityText: string;
  Backorderable: boolean;
  BrandID: number;
  BrandImageUrl: null;
  BrandName: string;
  BrandUrl: string;
  CatalogItemBehavior: string;
  CatalogItemTypeID: number;
  CategoryItems: CategoryItem[];
  ClickUrl: null;
  Condition: string;
  CurrentSellerID: number;
  CurrentSkuID: number;
  Descriptions: LinxMetadata[];
  DisplayCondition: boolean;
  DisplayPrice: string;
  DisplayStockQuantity: boolean;
  Edit: null;
  EstimatedReorderDate: null;
  ExtendedMetadatas: LinxMetadata[];
  Flags: unknown[];
  HasAnyItemDeliverable: boolean;
  HasCurrentSkuID: boolean;
  HasEstimatedReorderDate: boolean;
  HasPickup: boolean;
  IntegrationID: string;
  IsDeliverable: boolean;
  IsFreeShipping: boolean;
  IsGiftCertificate: boolean;
  IsInventoryAvailable: boolean;
  IsNew: boolean;
  IsPromotion: boolean;
  IsPurchasable: boolean;
  IsUndeliverable: boolean;
  IsUponRequest: boolean;
  Items: Item[];
  JoinCollection: null;
  LeafCategoryId: number;
  LeafCategoryName: string;
  LeafCategoryUrl: string;
  ListPrice: number;
  LoyaltyProgramPoints: unknown[];
  MaximumQtyAllowed: number;
  MediaGroups: MediaGroup[];
  Medias: Media[];
  MediaSmall: string;
  MetaDescription: string;
  MetaKeywords: string;
  MinimumQtyAllowed: number;
  MininumOfDistinctSKUs: number;
  MininumOfSKUsQuantity: number;
  MinLeadTimeDays: number;
  Name: string;
  Navigation: NavigationInfo[];
  NewFrom: null;
  NewTo: null;
  Options: Option[];
  OutOfStockFrom: null;
  OutOfStockTo: null;
  PageTitle: string;
  Preorderable: boolean;
  PreorderDate: string;
  Price: Price;
  PriceDescription: string;
  ProductDefinition: ProductDefinition;
  ProductDefinitionID: number;
  ProductID: number;
  ProductSelection: ProductSelectionItem;
  ProductTypeID: number;
  PromotionFrom: null;
  PromotionOnlyCheckout: boolean;
  PromotionPrice: number;
  PromotionTo: null;
  PropertyPath: string;
  PurchasingFormID: number;
  QuantityReturned: null;
  RatingAverage: number;
  RatingCount: number;
  ReplenishmentStatus: string;
  RetailPrice: number;
  RetailPriceMax: number;
  RetailPriceMin: number;
  ReviewCount: number;
  RootCategoryName: string;
  RootCategoryUrl: string;
  ShortDescription: string;
  SKU: string;
  StockBalance: string;
  StockOnHand: string;
  Suppliers: unknown[];
  Tags: Tag[];
  Tax: number;
  TaxationAmount: number;
  Uom: string;
  Url: string;

  VisibleFrom: string;
  VisibleTo: null;
}

export interface Item {
  Availability: string;
  AvailabilityText: string;
  Backorderable: boolean;
  BackorderLimit: null;
  BundleAdjustType: string;
  BundleAdjustValue: number;
  BundleGroup: null;
  BundleHierarchyPrice: null;
  BundleKitDiscount: string;
  BundleKitDiscountValue: number;
  BundlePriceType: string;
  BundleQuantity: number;
  BundleType: string;
  BuyBox: null;
  CatalogItemBehavior: string;
  CatalogItemTypeID: number;
  ComposeMinPrice: boolean;
  Depth: number;
  Edit: null;
  EstimatedReorderDate: null;
  ExtendedMetadatas: LinxMetadata[];
  HandlingDays: number;
  HasEstimatedReorderDate: boolean;
  Height: number;
  InStockHandlingDays: number;
  IntegrationID: string;
  IsBundleItemRequired: boolean;
  IsDeliverable: boolean;
  IsInventoryAvailable: boolean;
  IsPromotion: boolean;
  IsPurchasable: boolean;
  Items: Item[];
  ListPrice: number;
  LoyaltyProgramPoints: unknown[];
  MaximumQtyAllowed: number;
  MinimumQtyAllowed: number;
  Name: string;
  Options: Option[];
  Order: number;
  OutOfStockFrom: null;
  OutOfStockHandlingDays: number;
  OutOfStockTo: null;
  Preorderable: boolean;
  PreorderDate: string;
  Price: Price | null;
  PriceDescription: string;
  ProductDefinition: null;
  ProductID: number;
  ProductLevel: number;
  ProductPath: string;
  ProductTypeID: number;
  PromotionFrom: null;
  PromotionPrice: number;
  PromotionTo: null;
  ReplenishmentStatus: string;
  RetailPrice: number;
  SKU: string;
  SKUOptions: LinxMetadata[];
  StockBalance: number;
  Tax: number;
  TaxationAmount: number;
  UPC: null;
  UrlFriendly: string;

  VariationLevel: number;
  VariationPath: string;
  Weight: number;
  Width: number;
}

export interface Option {
  HasVariationLayout: boolean;
  IsVerificationRequired: boolean;
  Label: string;
  PropertyMetadataID: number;
  PropertyName: string;
  Values: PropertyItem[];
  VariationLayoutTemplate: string;
}

export interface Price {
  BestInstallment: MaxInstallmentsNoInterest | null;
  BestInstallmentCreditCard: null;
  BuyBox: SellerInfo;
  CurrentPaymentTerm: null;
  HasBuyBox: boolean;
  HasManySkus: boolean;
  HasSalesPrice: boolean;
  IsNullPrice: boolean;
  ItemAddFreeItem: null;
  ListPrice: number;
  ListPriceWithTax: number;
  MaxInstallmentsNoInterest: MaxInstallmentsNoInterest;
  MaxInstallmentsWithInterest: null;
  SalesPrice: number;
  SalesPriceDiscount: number;
  SalesPriceWithTax: number;
  SalesPriceWithTaxDiscount: number;
  Tax: number;
  TaxationAmount: number;
}

export interface MaxInstallmentsNoInterest {
  Discount: number;
  DiscountType: string;
  InstallmentPrice: number;
  Installments: number;
  InterestRate: number;
  PaymentName: string;
  PaymentType: string;
  RetailPrice: number;
  YearInterestRate: number;
}

export interface MediaGroup {
  Custom: null;
  IsPurchasable: boolean;
  Large: Media;
  Medias: Media[];
  MediaType: string;
  Medium: Media;
  ProductLevel: number;
  ProductPath: string;
  Small: Media;
  Thumbnail: Media;
  VariationLevel: number;
  VariationPath: string;
  Zoom: Media;
}

export interface Media {
  Attributes: unknown[];
  HasMediaPath: boolean;
  Height: number;
  Index: number;
  MaxHeight: number;
  MaxWidth: number;
  MediaPath: string;
  MediaSizeType: string;
  MediaType: string;
  ProductPath: string;
  Title: null;
  Url: null;
  VariationPath: string;
  Width: number;
}

export interface PageInfo {
  AlternateTitle: null;
  BodyClass: null;
  CanonicalLink: string;
  CanonicalNextLink: string;
  CanonicalPrevLink: string;
  DefaultAlternateTitle: null;
  DefaultDescription: null;
  DefaultMetaDescription: string;
  DefaultMetaKeywords: string;
  DefaultPageTitle: string;
  Description: string;
  HttpStatusCode: number;
  IndexPage: boolean;
  MetaDescription: string;
  MetaKeywords: string;
  PageTitle: string;
  RouteClass: string;
  SectionClass: null;
}

export interface Response {
  Errors: LinxError[];
  IsValid: boolean;
  RefreshBasket: boolean;
  ResponseCallBack: ResponseCallBack;
  SuccessMessage: null;
  Url: null;
  Warnings: unknown[];
}