// linx/utils/types/gridProductsJSON.ts
export interface WebPage {
  Response: Response;
  Model: Model;
  PageInfo: PageInfo;
}

export interface Model {
  Category: Category;
  CategoryID: number;
  Grid: Grid;
  GroupName: string;
  Message: null;
  Name: string;
  Navigation: NavigationInfo[];
  Provider: string;
  SearchId: string;
  SortAlias: null;
  Url: string;
}

export interface Category {
  CreatedBy: string;
  CreatedDate: string;
  IntegrationID: null;
  ModifiedBy: null;
  ModifiedDate: null;
}

export interface Grid {
  Facets: Facet[];
  HasNextPage: boolean;
  HasPreviousPage: boolean;
  HasProducts: boolean;
  PageCount: number;
  PageFirst: number;
  PageIndex: number;
  PageLast: number;
  PageNumber: number;
  PageSize: number;
  ProductCount: number;
  Products: Product[];
  ProviderCapabilities: ProviderCapabilities;
  Query: string;
  SortOptions: SortOption[];
  Spell: Spell;
  Terms: unknown[];
}
// ...
export interface PageInfo {
  AlternateTitle: string;
  BodyClass: string;
  CanonicalLink: string;
  CanonicalNextLink: string;
  CanonicalPrevLink: string;
  DefaultAlternateTitle: null;
  DefaultDescription: null;
  DefaultMetaDescription: string;
  DefaultMetaKeywords: string;
  DefaultPageTitle: string;
  Description: string;
  HttpStatusCode: number;
  IndexPage: boolean;
  MetaDescription: string;
  MetaKeywords: string;
  PageTitle: string;
  RouteClass: string;
  SectionClass: string;
}

export interface Response {
  Errors: LinxError[];
  IsValid: boolean;
  RefreshBasket: boolean;
  ResponseCallBack: ResponseCallBack;
  SuccessMessage: null;
  Url: null;
  Warnings: unknown[];
}