// linx/utils/types/auctionJSON.ts
export interface WebPage {
  Response: Response;
  Model: Model;
  PageInfo: PageInfo;
  PageParameters: PageParameters;
}

export interface Model {
  Grid: Grid;
  Message: null;
  Navigation: NavigationInfo[];
  ProductAuctions: ProductAuction[];
  SortAlias: null;
}

export interface Grid {
  Facets: Facet[];
  HasNextPage: boolean;
  HasPreviousPage: boolean;
  HasProducts: boolean;
  PageCount: number;
  PageFirst: number;
  PageIndex: number;
  PageLast: number;
  PageNumber: number;
  PageSize: number;
  ProductCount: number;
  Products: Product[];
  ProviderCapabilities: ProviderCapabilities;
  Query: string;
  SortOptions: SortOption[];
  Spell: Spell;
  Terms: unknown[];
}
// ...
export interface PageParameters {
}
// ...
export interface ProductAuction {
  BidAmountMultiplier: number;
  ExecutionFrom: Date;
  ExecutionTo: Date;
  InitialAmount: number;
  IsExecuting: boolean;
  MarketAmount: number;
  MarketAmountMultiplier: number;
  MinimunBidAmount: number;
  Name: string;
  ProductAuctionID: number;
  ProductID: number;
  ProductImage: string;
  ReservationMultiplier: number;
  SkuID: number;
  VisibleFrom: Date;
  VisibleTo: Date;
}

export interface PageInfo {
  AlternateTitle: null;
  BodyClass: string;
  CanonicalLink: string;
  CanonicalNextLink: string;
  CanonicalPrevLink: string;
  DefaultAlternateTitle: string;
  DefaultDescription: null;
  DefaultMetaDescription: string;
  DefaultMetaKeywords: string;
  DefaultPageTitle: string;
  Description: null;
  HttpStatusCode: number;
  IndexPage: boolean;
  MetaDescription: null;
  MetaKeywords: null;
  PageTitle: null;
  RouteClass: string;
  SectionClass: string;
}

export interface Response {
  Custom: PageParameters;
  Errors: LinxError[];
  IsValid: boolean;
  RefreshBasket: boolean;
  ResponseCallBack: ResponseCallBack;
  SuccessMessage: null;
  Url: null;
  Warnings: unknown[];
}