// linx/utils/types/auctionDetailJSON.ts
export interface WebPage {
  Response: Response;
  Model: Model;
  PageInfo: PageInfo;
  PageParameters: PageParameters;
}

export interface Model {
  BidAmountMultiplier: number;
  ExecutionFrom: Date;
  ExecutionTo: Date;
  InitialAmount: number;
  IsExecuting: boolean;
  MarketAmountMultiplier: number;
  MinimunBidAmount: number;
  Name: string;
  Navigation: NavigationInfo[];
  Product: Product;
  ProductAuctionID: number;
  ReservationAmount: number;
  VisibleFrom: Date;
  VisibleTo: Date;
}

export interface Product {
  AcceptanceTermID: number;
  AllowOnlyAdditionalPurchase: boolean;
  Associations: unknown[];
  Availability: string;
  AvailabilityText: string;
  Backorderable: boolean;
  BrandID: number;
  BrandImageUrl: null;
  BrandName: string;
  BrandUrl: string;
  CatalogItemBehavior: string;
  CatalogItemTypeID: number;
  CategoryItems: CategoryItem[];
  ClickUrl: null;
  Condition: string;
  CurrentSellerID: number;
  CurrentSkuID: number;
  Descriptions: LinxMetadata[];
  DisplayCondition: boolean;
  DisplayPrice: string;
  DisplayStockQuantity: boolean;
  Edit: null;
  EstimatedReorderDate: Date;
  ExtendedMetadatas: LinxMetadata[];
  Flags: unknown[];
  HasAnyItemDeliverable: boolean;
  HasCurrentSkuID: boolean;
  HasEstimatedReorderDate: boolean;
  HasPickup: boolean;
  IntegrationID: string;
  IsDeliverable: boolean;
  IsFreeShipping: boolean;
  IsGiftCertificate: boolean;
  IsInventoryAvailable: boolean;
  IsNew: boolean;
  IsPromotion: boolean;
  IsPurchasable: boolean;
  IsUndeliverable: boolean;
  IsUponRequest: boolean;
  Items: Item[];
  JoinCollection: null;
  LeafCategoryId: number;
  LeafCategoryName: string;
  LeafCategoryUrl: string;
  ListPrice: number;
  LoyaltyProgramPoints: unknown[];
  MaximumQtyAllowed: number;
  MediaGroups: MediaGroup[];
  Medias: Media[];
  MediaSmall: string;
  MetaDescription: string;
  MetaKeywords: string;
  MinimumQtyAllowed: number;
  MininumOfDistinctSKUs: number;
  MininumOfSKUsQuantity: number;
  MinLeadTimeDays: number;
  Name: string;
  Navigation: null;
  NewFrom: null;
  NewTo: null;
  Options: unknown[];
  OutOfStockFrom: null;
  OutOfStockTo: null;
  PageTitle: string;
  Preorderable: boolean;
  PreorderDate: Date;
  Price: Price;
  PriceDescription: string;
  ProductDefinition: ProductDefinition;
  ProductDefinitionID: number;
  ProductID: number;
  ProductSelection: ProductSelection;
  ProductTypeID: number;
  PromotionFrom: null;
  PromotionOnlyCheckout: boolean;
  PromotionPrice: number;
  PromotionTo: null;
  PropertyPath: string;
  PurchasingFormID: number;
  QuantityReturned: null;
  RatingAverage: number;
  RatingCount: number;
  ReplenishmentStatus: string;
  RetailPrice: number;
  RetailPriceMax: number;
  RetailPriceMin: number;
  ReviewCount: number;
  RootCategoryName: string;
  RootCategoryUrl: string;
  ShortDescription: string;
  SKU: string;
  StockBalance: string;
  StockOnHand: string;
  Suppliers: unknown[];
  Tags: unknown[];
  Tax: number;
  TaxationAmount: number;
  Uom: string;
  Url: string;
  UrlHelper: PageParameters;
  VisibleFrom: Date;
  VisibleTo: null;
}
// ...
export interface PageParameters {
}
// ...
export interface PageInfo {
  AlternateTitle: null;
  BodyClass: string;
  CanonicalLink: string;
  CanonicalNextLink: string;
  CanonicalPrevLink: string;
  DefaultAlternateTitle: string;
  DefaultDescription: null;
  DefaultMetaDescription: string;
  DefaultMetaKeywords: string;
  DefaultPageTitle: string;
  Description: null;
  HttpStatusCode: number;
  IndexPage: boolean;
  MetaDescription: null;
  MetaKeywords: null;
  PageTitle: null;
  RouteClass: string;
  SectionClass: null;
}

export interface Response {
  Custom: PageParameters;
  Errors: LinxError[];
  IsValid: boolean;
  RefreshBasket: boolean;
  ResponseCallBack: ResponseCallBack;
  SuccessMessage: null;
  Url: null;
  Warnings: unknown[];
}